Breakout `compileClassifier()`
